pyproject.toml
[project]
name = "surgcast"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "numpy",
    "h5py",
    "pyyaml",
    "openpyxl",
    "pandas",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["surgcast*"]
